requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
mysql-connector-python==8.1.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
matplotlib==3.7.2
Werkzeug==2.3.7